backend/requirements.txt
django
djangorestframework
python-dotenv
dj-database-url
psycopg2-binary
django_filter
django-cors-headers
pytest
pytest-django
model_bakery
django_extensions